Rename smart to containers
