+++ src/cow-react/modules/swap/containers/RemoveRecipient/index.tsx
