+++ src/cow-react/modules/swap/containers/TradeBasicDetails/index.tsx
