+++ src/cow-react/modules/swap/pure/SwapButton/SwapButton.tsx
@@ -10,7 +10,7 @@ import { GreyCard } from 'components/Card'
 import { GpEther } from 'constants/tokens'
 import { SupportedChainId } from 'constants/chains'
 import { AutoColumn } from 'components/Column'
-import { ApproveButton, ApproveButtonProps } from 'cow-react/modules/swap/smart/ApproveButton'
+import { ApproveButton, ApproveButtonProps } from 'cow-react/modules/swap/containers/ApproveButton'
 import * as styledEl from './styled'
 import { HandleSwapCallback } from 'cow-react/modules/swap/hooks/useHandleSwap'
 import { isSwapButtonPropsEqual } from 'cow-react/pages/NewSwap/propsChecker'
