+++ src/cow-react/modules/swap/pure/TradeRates/index.tsx
@@ -10,7 +10,7 @@ import { INITIAL_ALLOWED_SLIPPAGE_PERCENT } from 'constants/index'
 import { RowSlippage } from 'components/swap/TradeSummary/RowSlippage'
 import { Currency, CurrencyAmount, Percent } from '@uniswap/sdk-core'
 import { genericPropsChecker } from 'cow-react/pages/NewSwap/propsChecker'
-import { TradeBasicDetails } from 'cow-react/modules/swap/smart/TradeBasicDetails'
+import { TradeBasicDetails } from 'cow-react/modules/swap/containers/TradeBasicDetails'
 
 const SUBSIDY_INFO_MESSAGE_EXTENDED =
   SUBSIDY_INFO_MESSAGE + '. Click on the discount button on the right for more info.'
