+++ src/custom/pages/Swap/SwapMod.tsx
@@ -42,19 +42,22 @@ import { useErrorMessage } from 'hooks/useErrorMessageAndModal'
 import CowSubsidyModal from 'components/CowSubsidyModal'
 import { AlertWrapper } from './styleds'
 import { setMaxSellTokensAnalytics } from 'utils/analytics'
-import { ImportTokenModal } from 'cow-react/modules/swap/smart/ImportTokenModal'
+import { ImportTokenModal } from 'cow-react/modules/swap/containers/ImportTokenModal'
 import { CompatibilityIssuesWarning } from 'cow-react/modules/swap/pure/CompatibilityIssuesWarning'
-import { ConfirmSwapModalSetup, ConfirmSwapModalSetupProps } from 'cow-react/modules/swap/smart/ConfirmSwapModalSetup'
+import {
+  ConfirmSwapModalSetup,
+  ConfirmSwapModalSetupProps,
+} from 'cow-react/modules/swap/containers/ConfirmSwapModalSetup'
 import { useAtomValue } from 'jotai/utils'
 import { swapConfirmAtom } from 'cow-react/modules/swap/state/swapConfirmAtom'
 import { SwapButton, SwapButtonContext } from 'cow-react/modules/swap/pure/SwapButton/SwapButton'
-import { RemoveRecipient } from 'cow-react/modules/swap/smart/RemoveRecipient'
+import { RemoveRecipient } from 'cow-react/modules/swap/containers/RemoveRecipient'
 import { Price } from 'cow-react/modules/swap/pure/Price'
-import { TradeBasicDetails } from 'cow-react/modules/swap/smart/TradeBasicDetails'
+import { TradeBasicDetails } from 'cow-react/modules/swap/containers/TradeBasicDetails'
 import { BottomGrouping } from 'cow-react/modules/swap/pure/styled'
 import { ArrowWrapperLoader } from 'components/ArrowWrapperLoader'
 import { HighFeeWarning, NoImpactWarning } from 'components/SwapWarnings'
-import { FeesDiscount } from 'cow-react/modules/swap/smart/FeesDiscount'
+import { FeesDiscount } from 'cow-react/modules/swap/containers/FeesDiscount'
 import { RouteComponentProps } from 'react-router-dom'
 import EthFlowModal from 'components/swap/EthFlow'
 import { useSwapButtonContext } from 'cow-react/modules/swap/hooks/useSwapButtonContext'
